button with animated opacity
