+++ lib/src/features/onboarding_performance/view/widgets/app_icon_button.dart
@@ -71,6 +71,64 @@ class AppIconButton extends StatelessWidget {
   }
 }
 
+class AppButton extends StatelessWidget {
+  final String title;
+  final void Function() onTap;
+  final Color textColor;
+  final Color backgroundColor;
+  final Color borderColor;
+
+  const AppButton({
+    super.key,
+    required this.title,
+    required this.onTap,
+    required this.textColor,
+    required this.backgroundColor,
+    required this.borderColor,
+  });
+  const AppButton.purpleButton({
+    super.key,
+    required this.title,
+    required this.onTap,
+  })  : textColor = AppColor.whiteText,
+        backgroundColor = AppColor.purplePrimary,
+        borderColor = AppColor.purplePrimary;
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      onPressed: onTap,
+      style: ButtonStyle(
+        backgroundColor: MaterialStateProperty.all(backgroundColor),
+        elevation: MaterialStateProperty.all(0),
+        shape: MaterialStateProperty.all(
+          RoundedRectangleBorder(
+            side: BorderSide(color: borderColor, width: 2),
+            borderRadius: BorderRadius.circular(12),
+          ),
+        ),
+        minimumSize: MaterialStateProperty.all(
+          Size(
+            MediaQuery.of(context).size.width * 0.85,
+            MediaQuery.of(context).size.height * 0.06,
+          ),
+        ),
+      ),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(
+            title,
+            style: Theme.of(context)
+                .textTheme
+                .bodyLarge
+                ?.copyWith(color: textColor),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
 class OnboardControllButton extends StatelessWidget {
   final String titlePurple;
   final void Function() onTapPurple;
